+++ src/models/restful-model-collection.ts
@@ -10,7 +10,6 @@ export default class RestfulModelCollection<
   T extends RestfulModel
 > extends ModelCollection<any> {
   modelClass: typeof RestfulModel;
-  baseUrl?: string;
 
   constructor(modelClass: typeof RestfulModel, connection: NylasConnection) {
     super(modelClass, connection, modelClass.collectionName);